+++ crates/par-core/src/frontend_impl/types/checking.rs
@@ -6,8 +6,8 @@ use super::error::TypeError;
 use super::lattice::union_types;
 use super::{Context, TypeDefs};
 use crate::location::Span;
-use crate::par::types::implicit::infer_holes;
-use crate::par::types::lattice::intersect_types;
+use crate::frontend_impl::types::implicit::infer_holes;
+use crate::frontend_impl::types::lattice::intersect_types;
 use indexmap::{IndexMap, IndexSet};
 use std::collections::BTreeMap;
 use std::sync::Arc;
