+++ crates/par-core/src/frontend_impl/types/core.rs
@@ -1,7 +1,7 @@
 use super::super::language::{GlobalName, LocalName};
 use crate::location::{Span, Spanning};
-use crate::par::types::visit::Polarity;
-use crate::par::types::{TypeDefs, TypeError, visit};
+use crate::frontend_impl::types::visit::Polarity;
+use crate::frontend_impl::types::{TypeDefs, TypeError, visit};
 use arcstr::ArcStr;
 use im::HashSet;
 use std::collections::BTreeMap;
