+++ crates/par-core/src/frontend_impl/types/definitions.rs
@@ -1,6 +1,6 @@
 use crate::location::Span;
-use crate::par::language::{GlobalName, LocalName};
-use crate::par::types::{Type, TypeError, visit};
+use crate::frontend_impl::language::{GlobalName, LocalName};
+use crate::frontend_impl::types::{Type, TypeError, visit};
 use indexmap::{IndexMap, IndexSet};
 use std::sync::Arc;
