+++ crates/par-core/src/frontend_impl/types/tests.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 mod tests {
-    use crate::par::types::{Type, TypeDefs};
+    use crate::frontend_impl::types::{Type, TypeDefs};
 
     #[test]
     fn test_iterative_box_choice() {
@@ -14,11 +14,11 @@ mod tests {
                 Type::Box(_, inner) => match inner.as_ref() {
                     Type::Choice(_, branches) => {
                         assert_eq!(branches.len(), 2);
-                        assert!(branches.contains_key(&crate::par::language::LocalName {
+                        assert!(branches.contains_key(&crate::frontend_impl::language::LocalName {
                             span: crate::location::Span::None,
                             string: arcstr::ArcStr::from("method1"),
                         }));
-                        assert!(branches.contains_key(&crate::par::language::LocalName {
+                        assert!(branches.contains_key(&crate::frontend_impl::language::LocalName {
                             span: crate::location::Span::None,
                             string: arcstr::ArcStr::from("method2"),
                         }));
