+++ crates/par-core/src/runtime/compiler.rs
@@ -1,7 +1,7 @@
 use futures::{future::RemoteHandle, task::SpawnExt};
 
 use crate::{
-    par::{language::GlobalName, types::Type},
+    frontend_impl::{language::GlobalName, types::Type},
     runtime_impl::{
         flat::{stats::Rewrites, transpiler::Transpiled},
         readback::Handle,
@@ -25,7 +25,7 @@ impl Display for Compiled {
 
 impl Compiled {
     pub fn compile_file(
-        module: &crate::par::program::CheckedModule,
+        module: &crate::frontend_impl::program::CheckedModule,
         max_interactions: u32,
     ) -> Result<Self, crate::runtime_impl::RuntimeCompilerError> {
         Ok(Self {
