+++ crates/par-core/src/runtime/readback.rs
@@ -1,7 +1,7 @@
 use crate::location::Span;
-use crate::par::language::LocalName;
-use crate::par::primitive::{ParString, Primitive};
-use crate::par::types::{PrimitiveType, Type, TypeDefs};
+use crate::frontend_impl::language::LocalName;
+use crate::frontend_impl::primitive::{ParString, Primitive};
+use crate::frontend_impl::types::{PrimitiveType, Type, TypeDefs};
 use arcstr::ArcStr;
 use bytes::Bytes;
 use num_bigint::BigInt;
